Fixed drivetrain bug
Fixed the major bug that was causing power to drop as rpm increased. Drivetrain.getPower() is supposed to use the point-slope equation for lines to interpolate between two points on the power curve, but I wrote it y=(x1-x)+y1 rather than y=(x-x1)+y1, which is what it's supposed to be. Oops. 
Also cleaned up Drivetrain.setRPM() and changed the to-do.